backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.42.0
httpx==0.27.2
beautifulsoup4==4.12.3
supabase==2.11.0
python-dotenv==1.0.1
cryptography==44.0.0
